refactor: 게시물 파일 위치 변경 및 구조 개선
