+++ src/main/java/com/blog/domain/post/application/dto/PostCreateDTO.java
@@ -1,0 +1,4 @@
+package com.blog.domain.post.application.dto;
+
+public class PostCreateDTO {
+}
