+++ src/main/java/com/blog/domain/post/domain/repository/PostRepository.java
@@ -1,0 +1,4 @@
+package com.blog.domain.post.domain.repository;
+
+public interface PostRepository {
+}
