--- src/main/java/com/blog/domain/post/dto/PostCreateDTO.java
@@ -1,4 +1,0 @@
-package com.blog.domain.post.dto;
-
-public class PostCreateDTO {
-}
