--- src/main/java/com/blog/domain/post/repository/PostRepository.java
@@ -1,4 +1,0 @@
-package com.blog.domain.post.repository;
-
-public interface PostRepository {
-}
